requirements.txt
# FastAPI core
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.0

# Pydantic / Settings
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Object Storage & Image Processing
boto3==1.35.99
Pillow==11.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Dev / testing (optional but useful)
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1